Replacing electron-rebuild, with magicmirror-rebuild
To fix issues around errors like `An unhandled error occurred inside electron-rebuild
Could not detect abi for version 16.0.7 and runtime electron.  Updating "node-abi" might help solve this issue if it is a new release of electron

Error: Could not detect abi for version 16.0.7 and runtime electron.  Updating "node-abi" might help solve this issue if it is a new release of electron`